Extract dashboard page view
