+++ service-worker.js
@@ -61,6 +61,7 @@ const APP_SHELL = [
   '/js/client-list.js',
   '/js/nav.js',
   '/js/views-router.js',
+  '/js/dashboard-page-view.js',
   '/js/lens-pages.js',
   '/js/lens-page-shell.js',
   '/js/dashboard-widgets.js',
